Rename QueueBackend to MockNetworkBackend
Work on #1. I had a new idea for how this should work. It's pretty clear, but
basically I'll have a mock object and I'll say to expect something with
a certain payload on a certain address. Whenever I get a post or a get
I'll check it with the list and return whatever it says to return. Maybe
I'll have it check that everything was called too, I'm not sure.