Try to move forward on testing
It's going rather disastrously. I think I need to register an
InstanceCreator for gson to work right because I put interface
references in a bunch of classes instead of having concrete classes. So
that'll be fun. But I think there's only one call to get the floors, so
that could be good. We'll see what happens.